Added Primitive Death Indicator in BaseCardEffectsSlay Cards
Added logic in BaseUnitFunctions that turns the unit as disabled when it reaches 0 hp
Also an event is created in BaseUnitFunctions for triggereing stuff after death, nothing assigned to it yet
BaseCardEffects now has a function called SlayCheck. All Slay cards must implement this in their scripts, basically slay check is called after DealDamage. Slaycheck calls to targetObject if it is enabled or not, if not, a bool called isSlain is set to true. Slay cards will always have an isSlain if segment to activate next effect at slay