Added Card parameter in Card Effect Activate and addition of more cards
CardEffectActivate in BaseCardEffect are now accepting the Card SO of the card itself, this opens more possibilities for various card conditions. A new function is needed now for each card script, and that is AccessCard(Card), implement this before Actor Loading function
Addition of isPlayCounting  Logics. This is activated during ActorLoad function because it is always used in every card
Abilities rehaul because it aint working
Abilities in CardTagManager will be rethinked
There is a bug in Abilities where if an ability is used and a new ability card is played, it replaces its slot in the ability panel
Added the finals of Block and Counter synergy cards in game