Jigsaws can now be saved and loaded
Overhaul in saving cards, instead of just saving the Allcards of the enum. we now save it in CardAndJigsawWrapper which contains the card enum, jigsaw enum, and other jigsaw info. We then use that for loading. Cards without jigsaw has an AllCards input of "jigsaw" which can be thought of as null since the loading function will ignore all "jigsaw" entries